言語サーバー再起動コマンド、コマンド登録の処理を汎化
